// apps/main/src/models/v2/controlplane/moduleCatalogueEntry.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { ModuleCatalogueEntryTags } from './moduleCatalogueEntryTags';
import type { ModuleCatalogueStatus } from './moduleCatalogueStatus';

export interface ModuleCatalogueEntry {
  org_id: string;
  uuid: string;
  slug: string;
  display_name: string;
  description: string;
  resource_type: string;
  tags: ModuleCatalogueEntryTags;
  status: ModuleCatalogueStatus;
  /** @minimum 1 */
  resource_version: number;
  current_default_version_uuid?: string;
  previous_default_version_uuid?: string;
  /** @minimum 0 */
  managed_default_generation: number;
  created_at: string;
  archived_at?: string;
  archived_by?: string;
  archive_reason?: string;
}